+++ src/pages/checkout/index.js
@@ -1,0 +1,3 @@
+import Checkout from './checkout';
+
+export default Checkout;
